Working on different Ta calculation approaches
Moved t_noon and solar zenith (vs) calculations inside TSEB function.  Now the inputs to TSEB are the datetime (ee.Date) and the lon and lat as images.  The time could eventually be pulled off of one of the images and the lat/lon could be declared inside TSEB but they needed to be function parameters to keep the tests the same (for now).

Made separate methods for each of the air temperature calculation approaches and made them not lazy properties.  Removed ta_values and ta_source parameters from the Image class for now and instead these need to be set for each Ta method.  Added a ta_img parameter to the "et" method.