tech team meeting updates
Renamed all "etr" and "etf" bands, parameters, etc to "et_reference" and "et_fraction".
Removed output_type parameter from interpolate.
Added fix for when aggregate collection is empty.

The image value tests aren't all passing but I think this is a result of the changes to vpd.   I still need to add tests for all of the CFSR sources.