Initial attempt at reading sharpened TIR source
This is a kind of silly way of computing the LST from the TIR image.  I would prefer to call the Landsat class lst (or emissivity) functions instead of duplicating that code in disalexi.py, but this works for now.  We could also compute LST when the sharpened thermal images are built and either save that as the only band or save it as an extra band.