Tair calculation updates
Switched the Ta calculation functions so that the range of Ta values is built around a "central" value instead of being based on a "bottom" value.  This should make it a little more transparent what is happening when a user sets a given number of Ta steps.

Along with this changed the initial NLDAS code to use the maximum value for the day for the center target value, instead of using the first value in the day minus 40.  The result is almost identical (because 0 UTC was close to the max value in the western US) but should be a little more consistent.

Made the minimum ET value a parameter that can be set from the INI file with a default value of 0.01 to match previous versions of the code.  I think it should be set to 0 though for future runs since ALEXI can go to 0.

Changed the Ta selection to select either the average of the Ta values that bracket the minimum bias or the minimum (but increasing) bias if all the biases are positive.