Trying to get the tests to pass
Switched a lot of the test to use the C02 SR images since the default Ta and ALEXI is V003 and these were built from C02.

Removed the et_fraction tests.

The TSEB test isn't going to work since it is based on constant images but there is a reduceRegion call in the function now.  I'm not sure how to fix that.