Set stabil_iterations default to None
If stabil_iterations is set, the code will use that number.  If the value is not set, then it will be computed dynamically.  We need a way to set the value in order to test the function.

I also rewrote the dynamic a_pt calculations to remove all of the If calls.