Small changes to ta_coarse function
Rounding tair direct value to nearest tenth when used in generic ta_coarse call.  This could be to the nearest hundredth instead, but a little rounding should help with the compression and won't really matter since the Ta asset gets "smoothed" before use.

Returning the ta and bias bands used for interpolation.  This might help with identifying problematic calculations.