Changing most upper case variables to lower case
Changed the inputs to compute_G0.  Passing in water_mask instead of ndvi and albedo and passing in longitude and datetime instead of the computed time products which are only used in compute_G0 anyway.

Had to rename the psychrometric constant to from g to gamma to avoid interfering with G rename.

The USGS/NED asset is deprecated, so switched the test to use the SRTM one instead.