Don't allow Ta smoothing to fill holes, only replace existing values
This is only being applied to the new Ta mosaic assets, not the existing 1k and 10k steps assets.

Also fixed the geotransform in disalexi.py to match the transform of the ALEXI assets and other spots in the code.  Not sure why it was different.

Removed some of the Ta rounding calls.  These aren't needed since the computed Ta values are not being exported and we don't need to worry about storage.